les degres sont mis en Float

+++ app/src/main/java/com/celeste/thermco/DefineT.kt
@@ -22,14 +22,14 @@ class DefineT : AppCompatActivity() {
         type = intent.getIntExtra(EXTRA_SELECTOR, 0)
 
         sendBtn.setOnClickListener {
-
+            //TODO:faire que les valeur par defaut sont celle de la derniere fois pour les boutona
             if(degre_field.text.isEmpty() || hours_field.text.isEmpty() || duree_field.text.isEmpty() ){
                 Toast.makeText(this, "Tout remplir", Toast.LENGTH_LONG).show()
             }else{
                 val temporaire = Chauffage(
                     type,
                     day,
-                    degre_field.text.toString().toInt(),
+                    degre_field.text.toString().toFloat(),
                     hours_field.text.toString().toInt(),
                     duree_field.text.toString().toInt())
                 println(temporaire.toJSON().toString())
